refactor: use @raycast/utils getFavicon, consolidate color utils, add DRY helpers
