+++ extensions/folders/.context/plans/code-cleanup-refactoring-plan.md
@@ -1,0 +1,164 @@
+# Code Cleanup & Refactoring Plan
+
+## Executive Summary
+
+The codebase is in **excellent condition** with no dead code, unused imports, or TODO/FIXME markers. The exploration identified optimization opportunities for long-term maintainability rather than critical issues.
+
+## Current State Assessment
+
+| Metric | Status |
+|--------|--------|
+| Dead code | ✅ None found |
+| Unused imports | ✅ None found |
+| TODO/FIXME comments | ✅ None found |
+| Type safety | ✅ Strong TypeScript usage |
+| Performance patterns | ✅ Good memoization & caching |
+| Code organization | ✅ Well-structured |
+
+## Identified Opportunities
+
+### Priority 1: Color Utilities Consolidation (Low Effort, High Impact)
+
+**Problem:** Color utilities split between two files
+- `css-colors.ts` - CSS color name mapping
+- `utils.ts` lines 296-329 - Color validation (`isValidHexColor`, `normalizeHexColor`)
+
+**Solution:** Move color validation functions from `utils.ts` to `css-colors.ts`
+
+**Files to modify:**
+- `/src/css-colors.ts` - Add functions
+- `/src/utils.ts` - Remove functions, update imports
+- `/src/folder-edit-form.tsx` - Update imports
+
+---
+
+### Priority 2: Extract Map Creation Helpers (Low Effort, DRY)
+
+**Problem:** Duplicate map creation logic
+```typescript
+// Appears twice in form-utils.ts (lines 163, 242)
+new Map(filterWebsites(existingItems).map((w) => [normalizeUrl(w.url), w]))
+```
+
+**Solution:** Create helper functions in `form-utils.ts`:
+```typescript
+export function createWebsiteUrlMap(items: FolderItem[])
+export function createFolderMapById(folders: Folder[])
+```
+
+**Files to modify:**
+- `/src/form-utils.ts` - Add helpers, refactor usages
+
+---
+
+### Priority 3: Extract Duplicate Separation Helper (Low Effort, DRY)
+
+**Problem:** Repeated pattern in `add-items-form.tsx`
+```typescript
+const duplicateApps = apps.filter((path) => existingAppPaths.has(path));
+const newApps = apps.filter((path) => !existingAppPaths.has(path));
+```
+
+**Solution:** Create generic utility:
+```typescript
+export function separateDuplicates<T>(items: T[], existingSet: Set<T>): { duplicates: T[]; new: T[] }
+```
+
+**Files to modify:**
+- `/src/form-utils.ts` - Add helper
+- `/src/components/add-items-form.tsx` - Use helper
+
+---
+
+### Future Consideration: Split utils.ts (Medium Effort)
+
+**Note:** Not recommended for this cleanup session - `utils.ts` at 588 lines is large but functional. This is better addressed when the project stabilizes.
+
+**Potential future split:**
+- `utils/sorting.ts` - Sorting functions
+- `utils/icons.ts` - Icon utilities
+- `utils/urls.ts` - URL collection/formatting
+- `utils/actions.ts` - Bulk open actions
+
+---
+
+## Recommended Approach
+
+Given the codebase is already in excellent condition, I recommend **Option A: Targeted Cleanup** focusing only on the highest-value, lowest-risk improvements.
+
+### Scope
+
+1. ✅ Color utilities consolidation
+2. ✅ Map creation helpers
+3. ✅ Duplicate separation helper
+4. ❌ Skip utils.ts split (defer to future)
+5. ❌ Skip form type extraction (minimal value)
+
+### Estimated Effort
+
+- Total time: ~45-60 minutes
+- Risk: Low (small, isolated changes)
+- Test coverage: TypeScript compilation + ESLint
+
+---
+
+## Implementation Steps
+
+### Step 1: Color Utilities Consolidation
+
+1. Add `isValidHexColor` and `normalizeHexColor` to `/src/css-colors.ts`
+2. Update `/src/css-colors.ts` exports
+3. Update `/src/utils.ts` to import from `css-colors.ts`
+4. Update `/src/folder-edit-form.tsx` imports
+5. Run TypeScript check
+
+### Step 2: Map Creation Helpers
+
+1. Add `createWebsiteUrlMap` function to `/src/form-utils.ts`
+2. Refactor duplicated code in `form-utils.ts` (lines 163, 242)
+3. Run TypeScript check
+
+### Step 3: Duplicate Separation Helper
+
+1. Add `separateDuplicates` generic function to `/src/form-utils.ts`
+2. Refactor `add-items-form.tsx` to use helper
+3. Run TypeScript check
+
+### Step 4: Verification
+
+1. Run full TypeScript compilation: `./node_modules/.bin/tsc --noEmit`
+2. Run ESLint on src: `./node_modules/.bin/eslint src --ext .ts,.tsx`
+3. Manual verification of functionality
+
+---
+
+## Files to Modify
+
+| File | Changes |
+|------|---------|
+| `/src/css-colors.ts` | Add `isValidHexColor`, `normalizeHexColor` |
+| `/src/utils.ts` | Remove color functions, add import |
+| `/src/folder-edit-form.tsx` | Update import path for color functions |
+| `/src/form-utils.ts` | Add `createWebsiteUrlMap`, `separateDuplicates` helpers |
+| `/src/components/add-items-form.tsx` | Use `separateDuplicates` helper |
+
+---
+
+## Verification Plan
+
+1. **TypeScript compilation** - Zero errors
+2. **ESLint** - Zero errors on src directory
+3. **Import validation** - All imports resolve correctly
+4. **No behavior changes** - Refactoring only, no functional changes
+
+---
+
+## Change Log Entry
+
+After completion, add to project changelog:
+```
+## [Refactor] - YYYY-MM-DD
+- Consolidated color utilities into css-colors.ts
+- Added createWebsiteUrlMap helper to reduce duplication
+- Added separateDuplicates generic helper for DRY principle
+```
