+++ extensions/folders/eslint.config.js
@@ -1,5 +1,5 @@
+const { defineConfig } = require("eslint/config");
 const raycastConfig = require("@raycast/eslint-config");
 
-// Flatten the config array (ESLint 9 doesn't support nested arrays)
-module.exports = raycastConfig.flat();
+module.exports = defineConfig([...raycastConfig]);
